+++ packages/catalog/__tests__/core/utilities/manage-catalog-url/catalog-url-manager.test.ts
@@ -12,7 +12,7 @@ const setup = ({
   displayType?: GlobalTypes.ContentItemDisplayType;
 }): CatalogURLManager => {
   const parsedUrl = {
-    pathname: '/base',
+    pathName: '/base',
     searchString
   };
   const manager = new CatalogURLManager(parsedUrl);
@@ -25,7 +25,7 @@ const setup = ({
 
 describe('@thoughtindustries/catalog/core/CatalogURLManager', () => {
   it('should create new instance', () => {
-    const urlManager = new CatalogURLManager({ pathname: '/' });
+    const urlManager = new CatalogURLManager({ pathName: '/' });
     expect(urlManager).toBeInstanceOf(CatalogURLManager);
   });
 
@@ -205,15 +205,6 @@ describe('@thoughtindustries/catalog/core/CatalogURLManager', () => {
       const actual = manager.composeURLForSetPage(2);
       expect(actual).toMatchInlineSnapshot(`"/base?token=abc&page=2"`);
     });
-
-    it('should reset existing params and set page when is curated', () => {
-      const searchString =
-        '?labels=%5B%22label1%22%2C%22label2%22%5D&values=%5B%22value1%22%2C%22value2%22%5D&page=3&token=abc&query=test&content-types=%5B%22type1%22%5D&display-type=calendar&sort=createdAt%3Adesc';
-      const manager = setup({ searchString, isCurated: true });
-
-      const actual = manager.composeURLForSetPage(2);
-      expect(actual).toMatchInlineSnapshot(`"/base?page=2"`);
-    });
   });
 
   describe('composeURLForRemoveToken', () => {
@@ -316,14 +307,27 @@ describe('@thoughtindustries/catalog/core/CatalogURLManager', () => {
     });
   });
 
-  describe('composeURLForSetSearchTermForm', () => {
-    it('should set form url', () => {
+  describe('composeActionURLForSetSearchTermForm', () => {
+    it('should set form action url', () => {
       const searchString =
         '?labels=%5B%22label1%22%2C%22label2%22%5D&values=%5B%22value1%22%2C%22value2%22%5D&page=3&token=abc&query=test&content-types=%5B%22type1%22%5D&display-type=calendar&sort=createdAt%3Adesc';
       const manager = setup({ searchString });
 
-      const actual = manager.composeURLForSetSearchTermForm();
+      const actual = manager.composeActionURLForSetSearchTermForm();
       expect(actual).toMatchInlineSnapshot(`"/base"`);
+    });
+  });
+
+  describe('composeURLForSetSearchTermForm', () => {
+    it('should set search term', () => {
+      const searchString =
+        '?labels=%5B%22label1%22%2C%22label2%22%5D&values=%5B%22value1%22%2C%22value2%22%5D&page=3&token=abc&query=test&content-types=%5B%22type1%22%5D&display-type=calendar&sort=createdAt%3Adesc';
+      const manager = setup({ searchString });
+
+      const actual = manager.composeURLForSetSearchTermForm('new search term');
+      expect(actual).toMatchInlineSnapshot(
+        `"/base?labels=%5B%22label1%22%2C%22label2%22%5D&values=%5B%22value1%22%2C%22value2%22%5D&page=3&token=abc&query=new+search+term&content-types=%5B%22type1%22%5D&display-type=calendar&sort=createdAt%3Adesc"`
+      );
     });
   });