+++ packages/pagination/src/pagination.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import clsx from 'clsx';
+import { twMerge } from 'tailwind-merge';
 import { DEFAULT_HIDE_PAGE_LIST, DEFAULT_PAGE, DEFAULT_PAGE_SIZE } from './constants';
 import { ArrowLeftIcon, ArrowRightIcon, DoubleArrowLeftIcon, DoubleArrowRightIcon } from './icons';
 import { PaginationProps } from './types';
@@ -10,7 +10,8 @@ const Pagination = ({
   total,
   pageSize = DEFAULT_PAGE_SIZE,
   getPageLink,
-  hidePageList = DEFAULT_HIDE_PAGE_LIST
+  hidePageList = DEFAULT_HIDE_PAGE_LIST,
+  linkComponent
 }: PaginationProps): JSX.Element => {
   // derived values
   const lastPage = Math.ceil(total / pageSize) || 1;
@@ -27,36 +28,37 @@ const Pagination = ({
   const disabledClassnames =
     'cursor-default pointer-events-none text-gray-400 bg-gray-300 border-gray-300';
   const enabledClassnames = 'text-gray-600';
-  const firstPageClassnames = clsx(
+  const firstPageClassnames = twMerge(
     [pageBaseClassnames, 'rounded rounded-r-none border-r-0'],
     !hasPrevPage ? disabledClassnames : enabledClassnames
   );
-  const prevPageClassnames = clsx(
+  const prevPageClassnames = twMerge(
     [pageBaseClassnames, 'rounded rounded-l-none border-r-1 mr-2'],
     !hasPrevPage ? disabledClassnames : enabledClassnames
   );
-  const nextPageClassnames = clsx(
+  const nextPageClassnames = twMerge(
     [pageBaseClassnames, 'rounded rounded-r-none border-r-0 ml-2'],
     !hasNextPage ? disabledClassnames : enabledClassnames
   );
-  const lastPageClassnames = clsx(
+  const lastPageClassnames = twMerge(
     [pageBaseClassnames, 'rounded rounded-l-none border-r-1'],
     !hasNextPage ? disabledClassnames : enabledClassnames
   );
 
   // components
+  const PaginationLinkComponent = linkComponent || 'a';
   const visiblePagesContent = !hidePageList
     ? visiblePages.map(({ number, label, isActive, isFirst, isLast }) => {
         const firstLastBorderClassnames = isFirst
           ? 'border-l-1 border-r-0 rounded-r-none'
           : 'border-r-1 border-l-0 rounded-l-none';
         const borderClassnames =
-          isFirst || isLast ? clsx('rounded', firstLastBorderClassnames) : 'border-x-0';
+          isFirst || isLast ? twMerge('rounded', firstLastBorderClassnames) : 'border-x-0';
         return (
-          <a
+          <PaginationLinkComponent
             key={`catalog-page-${label}`}
             href={getPageLink(number)}
-            className={clsx(
+            className={twMerge(
               [pageBaseClassnames, borderClassnames],
               isActive
                 ? 'cursor-default pointer-events-none bg-accent border-accent text-accent-contrast'
@@ -64,42 +66,52 @@ const Pagination = ({
             )}
           >
             {label}
-          </a>
+          </PaginationLinkComponent>
         );
       })
     : null;
 
   return (
     <div className="mx-2 my-4 flex flex-wrap-reverse items-center justify-between">
-      <div className="mt-2 flex items-center justify-start">
-        <span>
-          Showing<strong>{` ${start}-${end} `}</strong>of<strong>{` ${total} items`}</strong>
-        </span>
-      </div>
+      {start <= end && (
+        <div className="mt-2 flex items-center justify-start">
+          <span>
+            Showing<strong>{` ${start}-${end} `}</strong>of<strong>{` ${total} items`}</strong>
+          </span>
+        </div>
+      )}
       {!!visiblePages.length && (
         <div className="mt-2 flex items-center justify-end">
           <div className="flex justify-center">
-            <a href={getPageLink(1)} className={firstPageClassnames} aria-label="rewind">
+            <PaginationLinkComponent
+              href={getPageLink(1)}
+              className={firstPageClassnames}
+              aria-label="rewind"
+            >
               <DoubleArrowLeftIcon />
-            </a>
-            <a
+            </PaginationLinkComponent>
+            <PaginationLinkComponent
               href={getPageLink(prevPage)}
               className={prevPageClassnames}
               aria-label="navigateleft"
             >
               <ArrowLeftIcon />
-            </a>
+            </PaginationLinkComponent>
             {visiblePagesContent}
-            <a
+            <PaginationLinkComponent
               href={getPageLink(nextPage)}
               className={nextPageClassnames}
               aria-label="navigateright"
             >
               <ArrowRightIcon />
-            </a>
-            <a href={getPageLink(lastPage)} className={lastPageClassnames} aria-label="fastforward">
+            </PaginationLinkComponent>
+            <PaginationLinkComponent
+              href={getPageLink(lastPage)}
+              className={lastPageClassnames}
+              aria-label="fastforward"
+            >
               <DoubleArrowRightIcon />
-            </a>
+            </PaginationLinkComponent>
           </div>
         </div>
       )}