+++ tooling/cli/lib/deploy.js
@@ -8,7 +8,7 @@ const childProcess = require('child_process');
 const tar = require('tar');
 const os = require('os');
 
-const { getFilePaths } = require('./helpers/filepaths');
+const { getFilePaths, getFileContents } = require('./helpers/filepaths');
 const { instanceEndpoint } = require('./helpers/urls');
 const { generateTranslationFile } = require('./file-generators');
 const {
@@ -18,6 +18,8 @@ const {
   hashQuery
 } = require('./helpers/graphql');
 const { dirHasAtoms, getAtomsHash, compileStyles } = require('./helpers/atoms');
+const { minimatch } = require('minimatch');
+const { DEFAULT_GRAPHQL_SOURCE_PATHS } = require('./helpers/constants');
 
 const OP_DIR = process.cwd();
 const TMP_DIR = os.tmpdir();
@@ -25,6 +27,7 @@ const configPath = path.join(OP_DIR, '/ti-config');
 const config = require(configPath);
 
 const INSTANCE_NAME = process.env.INSTANCE_NAME;
+const DEVELOPMENT_BUILD = process.env.DEVELOPMENT_BUILD;
 
 const KEY_QUERY = /* GraphQL */ `
   query CompanyDetailsQuery($nickname: String!) {
@@ -101,8 +104,13 @@ class DeploymentError extends Error {
 
   // hash queries for whitelisting
   await writeGraphqlManifest();
+
+  await maybeRenameEnvFile();
+
   await uploadHeliumProject(policyData);
-  // reset translations file to include all keys for local development
+
+  await maybeResetEnvFile();
+  // // reset translations file to include all keys for local development
   await resetTranslationFile();
   const batchJobId = await triggerBatch(instance, key);
 
@@ -137,10 +145,55 @@ process.on('message', message => {
   console.log('>>> Deploy process receive message', message);
 });
 
+async function maybeRenameEnvFile() {
+  // temporarily swap env files if it's a dev build.
+  // the CF Worker still loads an env file into process.env if it's present
+  // so we could end up with a project built with .env.development running
+  // w/ vars passed from .env instead.
+  if (isDevelopmentBuild()) {
+    await swapEnvFiles('.env', '.env.development', '.env.backup');
+  }
+
+  return;
+}
+
+async function maybeResetEnvFile() {
+  if (isDevelopmentBuild()) {
+    await swapEnvFiles('.env', '.env.backup', '.env.development');
+  }
+
+  return;
+}
+
+async function swapEnvFiles(target, source, newFilePath) {
+  const targetPath = path.join(OP_DIR, target);
+  const sourcePath = path.join(OP_DIR, source);
+
+  const hasTarget = await getFileContents(targetPath);
+  const hasSource = await getFileContents(sourcePath);
+
+  if (hasTarget && hasSource) {
+    const newPath = path.join(OP_DIR, newFilePath);
+    await rename(targetPath, newPath);
+    await rename(sourcePath, targetPath);
+  }
+
+  return;
+}
+
+function isDevelopmentBuild() {
+  return String(DEVELOPMENT_BUILD) === 'true';
+}
+
 async function buildProject(hasAtoms) {
   return new Promise((resolve, reject) => {
     const exec = childProcess.exec;
-    const buildCommandSuffix = hasAtoms ? 'atoms' : 'vite';
+
+    let buildCommandSuffix = hasAtoms ? 'atoms' : 'vite';
+    if (buildCommandSuffix === 'vite' && isDevelopmentBuild()) {
+      buildCommandSuffix = 'development';
+    }
+
     const parseProcess = exec(`npm run build:css && npm run build:${buildCommandSuffix}`);
 
     parseProcess.stdout.pipe(process.stdout);
@@ -179,13 +232,32 @@ async function writeGraphqlManifest() {
 }
 
 async function hashGraphqlQueries() {
-  const pagesFilePaths = await getFilePaths(path.join(OP_DIR, 'pages'));
-  const componentsFilePaths = await getFilePaths(path.join(OP_DIR, 'components'));
-  const tiFilepaths = await getFilePaths(path.join(OP_DIR, 'node_modules/@thoughtindustries'));
-  const filePaths = pagesFilePaths.concat(componentsFilePaths).concat(tiFilepaths);
+  let pathMatchPatterns = DEFAULT_GRAPHQL_SOURCE_PATHS;
+  const { content = [] } = config;
+
+  if (content.length) {
+    pathMatchPatterns = content;
+  }
+
+  const projectFilePaths = await getFilePaths(path.join(OP_DIR, '/'));
+  let sourcePaths = [];
+
+  pathMatchPatterns.forEach(matchPattern => {
+    const absoluteMatchPattern = path.join(OP_DIR, matchPattern);
+    const matches = minimatch.match(projectFilePaths, absoluteMatchPattern);
+    sourcePaths = sourcePaths.concat(matches);
+  });
+
+  const tiFilePaths = await getFilePaths(
+    path.join(OP_DIR, 'node_modules/@thoughtindustries'),
+    [],
+    false
+  );
+
+  sourcePaths = sourcePaths.concat(tiFilePaths);
 
   const queryHashMap = {};
-  const querySources = await gatherQuerySources(filePaths, '__tests__');
+  const querySources = await gatherQuerySources(sourcePaths, '__tests__');
 
   if (querySources.length > 0) {
     const fragmentMap = buildFragmentMap(querySources);