+++ tooling/cli/lib/helpers/constants.js
@@ -1,0 +1,6 @@
+module.exports = {
+  DEFAULT_GRAPHQL_SOURCE_PATHS: [
+    './pages/**/*.{js,jsx,ts,tsx}',
+    './components/**/*.{js,jsx,ts,tsx}'
+  ]
+};