+++ tooling/cli/lib/helpers/filepaths.js
@@ -1,7 +1,8 @@
 const { readdir, stat } = require('fs/promises');
 const path = require('path');
+const { access, readFile } = require('fs/promises');
 
-const getFilePaths = async (dir, filePaths = []) => {
+const getFilePaths = async (dir, filePaths = [], skipNodeModules = true) => {
   const newFilePaths = filePaths;
 
   try {
@@ -14,6 +15,12 @@ const getFilePaths = async (dir, filePaths = []) => {
       if (fileStat.isFile()) {
         newFilePaths.push(filePath);
       } else {
+        if (skipNodeModules) {
+          if (filePath.includes('node_modules') && !filePath.includes('@thoughtindustries')) {
+            continue;
+          }
+        }
+
         await getFilePaths(filePath, newFilePaths);
       }
     }
@@ -33,4 +40,14 @@ const filePathIsValid = filePath => {
   );
 };
 
-module.exports = { getFilePaths, filePathIsValid };
+const getFileContents = async path => {
+  try {
+    await access(path);
+    const contents = await readFile(path);
+    return contents;
+  } catch {
+    return null;
+  }
+};
+
+module.exports = { getFilePaths, filePathIsValid, getFileContents };