+++ tooling/template-base/components/Navigation/CurrentUserNavBar.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useState } from 'react';
 import Logo from '../Logo/Logo';
-import dropDownClosed from '../Assets/dropDownClosed';
-import dropDownOpen from '../Assets/dropDownOpen';
+import DropDownClosed from '../Assets/DropDownClosed';
+import DropDownOpen from '../Assets/DropDownOpen';
 import Avatar from '../Avatar/Avatar';
 
 export default function CurrentUserNavBar() {
@@ -37,7 +37,7 @@ export default function CurrentUserNavBar() {
             <button id="icon" type="submit" className="" onClick={() => setNavbar(!navbar)}>
               <div className="flex items-center space-x-3 mr-3">
                 <Avatar size="" />
-                {navbar ? <div>{dropDownOpen} </div> : <div>{dropDownClosed}</div>}
+                {navbar ? <div>{DropDownOpen} </div> : <div>{DropDownClosed}</div>}
               </div>
             </button>
           </div>