+++ packages/cart/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/cart@1.2.4...@thoughtindustries/cart@1.2.5) (2023-11-01)
+
+**Note:** Version bump only for package @thoughtindustries/cart
+
+
+
+
+
+## [1.2.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/cart@1.2.3...@thoughtindustries/cart@1.2.4) (2023-08-28)
+
+**Note:** Version bump only for package @thoughtindustries/cart
+
+
+
+
+
 ## [1.2.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/cart@1.2.2...@thoughtindustries/cart@1.2.3) (2023-07-19)
 
 **Note:** Version bump only for package @thoughtindustries/cart