+++ packages/video-player/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/video-player@1.1.4...@thoughtindustries/video-player@1.1.5) (2023-11-01)
+
+**Note:** Version bump only for package @thoughtindustries/video-player
+
+
+
+
+
+## [1.1.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/video-player@1.1.3...@thoughtindustries/video-player@1.1.4) (2023-08-28)
+
+**Note:** Version bump only for package @thoughtindustries/video-player
+
+
+
+
+
 ## [1.1.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/video-player@1.1.2...@thoughtindustries/video-player@1.1.3) (2023-07-19)
 
 **Note:** Version bump only for package @thoughtindustries/video-player