+++ tooling/template-base-essentials/CHANGELOG.md
@@ -3,6 +3,93 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.7...@thoughtindustries/helium-template-essentials@1.4.0) (2023-11-14)
+
+
+### Features
+
+* allow deployment of development builds ([#231](https://github.com/thoughtindustries/helium/issues/231)) ([ed06e65](https://github.com/thoughtindustries/helium/commit/ed06e65bd62b964f002bed03295ca6a6fcfbb5f9))
+
+
+
+
+
+# [1.3.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.3.0-alpha.2...@thoughtindustries/helium-template-essentials@1.3.0) (2023-11-09)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template-essentials
+
+
+
+
+
+# [1.3.0-alpha.2](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.7...@thoughtindustries/helium-template-essentials@1.3.0-alpha.2) (2023-11-09)
+
+
+### Features
+
+* allow deployment of development builds ([#231](https://github.com/thoughtindustries/helium/issues/231)) ([ed06e65](https://github.com/thoughtindustries/helium/commit/ed06e65bd62b964f002bed03295ca6a6fcfbb5f9))
+
+
+
+
+
+# [1.3.0-alpha.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.7...@thoughtindustries/helium-template-essentials@1.3.0-alpha.1) (2023-11-08)
+
+
+### Features
+
+* allow deployment of development builds ([#231](https://github.com/thoughtindustries/helium/issues/231)) ([ed06e65](https://github.com/thoughtindustries/helium/commit/ed06e65bd62b964f002bed03295ca6a6fcfbb5f9))
+
+
+
+
+
+# [1.3.0-alpha.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.7...@thoughtindustries/helium-template-essentials@1.3.0-alpha.0) (2023-11-07)
+
+
+### Features
+
+* allow deployment of development builds ([#231](https://github.com/thoughtindustries/helium/issues/231)) ([ed06e65](https://github.com/thoughtindustries/helium/commit/ed06e65bd62b964f002bed03295ca6a6fcfbb5f9))
+
+
+
+
+
+## [1.2.7](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.6...@thoughtindustries/helium-template-essentials@1.2.7) (2023-11-01)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template-essentials
+
+
+
+
+
+## [1.2.6](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.5...@thoughtindustries/helium-template-essentials@1.2.6) (2023-09-28)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template-essentials
+
+
+
+
+
+## [1.2.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.4...@thoughtindustries/helium-template-essentials@1.2.5) (2023-08-28)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template-essentials
+
+
+
+
+
+## [1.2.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.3...@thoughtindustries/helium-template-essentials@1.2.4) (2023-08-22)
+
+
+### Bug Fixes
+
+* helium init fixes ([#224](https://github.com/thoughtindustries/helium/issues/224)) ([12248ec](https://github.com/thoughtindustries/helium/commit/12248ec48a121e5c6b129e3810dd1a7ed9b01a7b))
+
+
+
+
+
 ## [1.2.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template-essentials@1.2.2...@thoughtindustries/helium-template-essentials@1.2.3) (2023-07-19)
 
 **Note:** Version bump only for package @thoughtindustries/helium-template-essentials